Move Home Component to home folder
