+++ src/App.jsx
@@ -1,7 +1,7 @@
 import {
   BrowserRouter, Routes, Route,
 } from 'react-router-dom';
-import Home from './components/Home';
+import Home from './components/Home/Home';
 import Navigation from './components/Navigation';
 import Login from './components/Login';
 import Dashboard from './components/Dashboard';
