+++ src/components/Home/Home.jsx
@@ -1,16 +1,9 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { Link } from 'react-router-dom';
-import { getProperties } from '../redux/property/propertySlice';
-import '../main.scss';
+import React from 'react';
+import SecondSection from './SecondSection';
+import '../../main.scss';
+
 
 const Home = () => {
-  const dispatch = useDispatch();
-  const newProperites = useSelector((state) => state.properties.properties);
-
-  useEffect(() => {
-    dispatch(getProperties());
-  }, [dispatch]);
   return (
     <div>
       <div className="BannerPart">
@@ -51,27 +44,7 @@ const Home = () => {
         </div>
       </div>
       <div>
-        <h3>Latest propertties</h3>
-        <div>
-          {newProperites.map((property) => (
-            <div key={property.id}>
-              <Link to={`/property-details/${property.id}`}>
-                <h4>{property.title}</h4>
-                <p>{property.price}</p>
-                <p>{property.address}</p>
-                <p>
-                  For
-                  {property.ownership_type.name}
-                </p>
-                <p>{property.property_type.name}</p>
-                <p>
-                  Posted  by :
-                  {property.created_by.email}
-                </p>
-              </Link>
-            </div>
-          ))}
-        </div>
+        <SecondSection />
       </div>
     </div>
   );
